= 1.3.1 = * Minor performance improvements * MU activation fix * Fix header already send warning * Fix PHP issue on plugin remove
git-svn-id: https://plugins.svn.wordpress.org/adblock-notify-by-bweb/trunk@1040820 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ adblock-notify.php
@@ -3,7 +3,7 @@
  * Plugin Name: Adblock Notify by b*web
  * Plugin URI: http://b-website.com/
  * Description: An Adblock detection and nofitication plugin with get around options and a lot of settings. Dashboard widget with adblock counter included!
- * Version: 1.3
+ * Version: 1.3.1
  * Author: Brice CAPOBIANCO
  * Author URI: b-website.com
  * Text Domain: an-translate
@@ -62,7 +62,7 @@ add_action('plugins_loaded', 'an_translate_load_textdomain', 1);
 /***************************************************************
  * Load plugin files
  ***************************************************************/
-require_once( AN_PATH . 'lib/titan-framework/titan-framework.php' );
+require_once( AN_PATH . 'lib/titan-framework/titan-framework-embedder.php' );
 
 $anFiles = array('options', 'functions', 'widget', 'files');
 foreach ($anFiles as $anFile) {
@@ -145,7 +145,6 @@ add_filter('user_contactmethods', 'user_contactmethods_example');
 /***************************************************************
  * Add settings link on plugin list page
  ***************************************************************/
-
 function an_settings_link($links) {
     $links[] = '<a href="options-general.php?page=' . AN_ID . '">' . __('Settings', 'an-translate') . '</a>';
     return $links;
@@ -157,25 +156,21 @@ add_filter('plugin_action_links_' . AN_BASE, 'an_settings_link');
 /***************************************************************
  * Add custom meta link on plugin list page
  ***************************************************************/
-if (!function_exists('an_meta_links')) {
-
-    function an_meta_links($links, $file) {
-        if (strpos($file, 'adblock-notify.php') !== false) {
-            $links[0] = '<a href="http://b-website.com/" target="_blank"><img src="' . AN_URL . 'img/icon-bweb.svg" style="margin-bottom: -4px; width: 18px;" alt="b*web"/></a>&nbsp;&nbsp;' . $links[0];
-            $links[] = '<a href="http://b-website.com/category/plugins" target="_blank" title="' . __('More b*web Plugins', 'an-translate') . '">' . __('More b*web Plugins', 'an-translate') . '</a>';
-            $links[] = '<a href="https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=7Z6YVM63739Y8" target="_blank" title="' . __('Donate', 'an-translate') . '"><strong>' . __('Donate', 'an-translate') . '</strong></a>';
-        }
-        return $links;
-    }
-
-    add_filter('plugin_row_meta', 'an_meta_links', 10, 2);
+function an_meta_links($links, $file) {
+	if (strpos($file, 'adblock-notify.php') !== false) {
+		$links[0] = '<a href="http://b-website.com/" target="_blank"><img src="' . AN_URL . 'img/icon-bweb.svg" style="margin-bottom: -4px; width: 18px;" alt="b*web"/></a>&nbsp;&nbsp;' . $links[0];
+		$links[] = '<a href="http://b-website.com/category/plugins" target="_blank" title="' . __('More b*web Plugins', 'an-translate') . '">' . __('More b*web Plugins', 'an-translate') . '</a>';
+		$links[] = '<a href="https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=7Z6YVM63739Y8" target="_blank" title="' . __('Donate', 'an-translate') . '"><strong>' . __('Donate', 'an-translate') . '</strong></a>';
+	}
+	return $links;
 }
+
+add_filter('plugin_row_meta', 'an_meta_links', 10, 2);
 
 
 /***************************************************************
  * Admin Panel Favico
  ***************************************************************/
-
 function an_add_favicon() {
     $screen = get_current_screen();
     if ($screen->id != 'toplevel_page_' . AN_ID)
@@ -191,7 +186,6 @@ add_action('admin_head', 'an_add_favicon');
 /***************************************************************
  * Create random selectors and files on plugin activation
  ***************************************************************/
-
 function adblocker_notify_activate() {
     add_action('tf_create_options', 'an_create_options');
     an_save_setting_random_selectors();
@@ -212,27 +206,40 @@ if (function_exists('register_uninstall_hook')) {
 
 //Remove directory
 function an_delete_temp_folder($dirPath) {
-    $files = glob($dirPath . '*', GLOB_MARK);
-    foreach ($files as $file) {
-        if (is_dir($file)) {
-            self::deleteDir($file);
-        } else {
-            unlink($file);
-        }
-    }
-    rmdir($dirPath);
+	
+	if( file_exists( $dirPath ) ) {
+  
+		$files = glob($dirPath . '*', GLOB_MARK);
+		foreach ($files as $file) {
+			if (is_dir($file)) {
+				self::deleteDir($file);
+			} else {
+				unlink($file);
+			}
+		}
+		
+			rmdir($dirPath);
+	}
+		
 }
 
 //Uninstall function
 function adblocker_notify_uninstall() {
+   
     // Remove temp files
     $anTempDir = unserialize(get_option('adblocker_notify_selectors'));
     an_delete_temp_folder($anTempDir['temp-path']);
+   
+    //Remove TitanFramework Generated Style
     $uploadDir = wp_upload_dir();
-    unlink(trailingslashit($uploadDir['basedir']) . 'titan-framework-adblocker_notify-css.css');
-
+    $TfCssFile = trailingslashit( $uploadDir['basedir']  ) . 'titan-framework-adblocker_notify-css.css';
+	
+	if( file_exists( $TfCssFile ) )
+    	unlink( $TfCssFile );
+		
     // Remove option from DB
     delete_option('adblocker_notify_options');
     delete_option('adblocker_notify_counter');
     delete_option('adblocker_notify_selectors');
+	
 }
